Day 4

+++ backend/scripts/cleanUnusedImages.js
@@ -12,7 +12,7 @@ import GeneralInfoModel from "../models/GeneralInfoModel.js";
 import ProductModel from "../models/ProductModel.js";
 import UserModel from "../models/UserModel.js";
 import BlogModel from "../models/BlogModel.js";
-
+import ProfileModel from "../models/ProfileModel.js";
 
 // Setup __dirname for ES module
 const __filename = fileURLToPath(import.meta.url);
@@ -27,7 +27,7 @@ await mongoose.connect(process.env.MONGO_URI, {
   useUnifiedTopology: true,
 });
 
-// Collect used images
+// Set to collect used image filenames
 const usedImages = new Set();
 const addImage = (img) => {
   if (img && typeof img === "string") {
@@ -36,35 +36,49 @@ const addImage = (img) => {
 };
 
 const collectUsedImages = async () => {
+  // Carousels
   const carousels = await CarouselModel.find({}, "imgSrc");
-  carousels.forEach((item) => addImage(item.imgSrc));
+  carousels.forEach(({ imgSrc }) => addImage(imgSrc));
 
+  // Feature images
   const features = await FeatureImageModel.find({}, "imgSrc");
-  features.forEach((item) => addImage(item.imgSrc));
+  features.forEach(({ imgSrc }) => addImage(imgSrc));
 
+  // General info logos
   const infos = await GeneralInfoModel.find({}, "PrimaryLogo SecondaryLogo Favicon");
-  infos.forEach((item) => {
-    addImage(item.PrimaryLogo);
-    addImage(item.SecondaryLogo);
-    addImage(item.Favicon);
+  infos.forEach(({ PrimaryLogo, SecondaryLogo, Favicon }) => {
+    addImage(PrimaryLogo);
+    addImage(SecondaryLogo);
+    addImage(Favicon);
   });
 
+  // Products
   const products = await ProductModel.find({}, "thumbnailImage images");
-  products.forEach((product) => {
-    addImage(product.thumbnailImage);
-    product.images?.forEach(addImage);
+  products.forEach(({ thumbnailImage, images }) => {
+    addImage(thumbnailImage);
+    images?.forEach(addImage);
   });
 
-  const users = await UserModel.find({}, "userImage");
-  users.forEach((user) => addImage(user.userImage));
+  // Profiles
+  const profiles = await ProfileModel.find({}, "profilePhoto galleryPhotos coverPhoto brandLogo");
+  profiles.forEach(({ profilePhoto, galleryPhotos, coverPhoto, brandLogo }) => {
+    addImage(profilePhoto);
+    addImage(coverPhoto);
+    addImage(brandLogo);
+    galleryPhotos?.forEach(addImage);
+  });
 
+  // Users
+  const users = await UserModel.find({}, "userImage");
+  users.forEach(({ userImage }) => addImage(userImage));
+
+  // Blogs
   const blogs = await BlogModel.find({}, "thumbnailImage");
-  blogs.forEach((user) => addImage(user.userImage));
+  blogs.forEach(({ thumbnailImage }) => addImage(thumbnailImage));
 };
 
 await collectUsedImages();
 
-// Delete unused images
 const uploadsDir = path.join(__dirname, "../uploads");
 
 fs.readdir(uploadsDir, (err, allFiles) => {
@@ -76,6 +90,11 @@ fs.readdir(uploadsDir, (err, allFiles) => {
   const unusedFiles = allFiles.filter((file) => !usedImages.has(file));
   let deletedCount = 0;
 
+  if (unusedFiles.length === 0) {
+    console.log("✅ No unused images to delete. Uploads folder is clean.");
+    process.exit(0);
+  }
+
   unusedFiles.forEach((file) => {
     const fullPath = path.join(uploadsDir, file);
     fs.unlink(fullPath, (err) => {
@@ -85,14 +104,9 @@ fs.readdir(uploadsDir, (err, allFiles) => {
         deletedCount++;
         if (deletedCount === unusedFiles.length) {
           console.log(`✅ Cleanup complete. ${deletedCount} unused image(s) deleted.`);
-          process.exit();
+          process.exit(0);
         }
       }
     });
   });
-
-  if (unusedFiles.length === 0) {
-    console.log("✅ No unused images to delete. Uploads folder is clean.");
-    process.exit();
-  }
 });
